Create app.R and Add Data
Created the app.R script (ui and server all-in-one) from final app script 'finalapp' that was formed when assessing better metrics for measuring housing affordability across the United States. This can be the base for measuring how much we spend and displaying key graphics we might want to interact with.