Review dataprep.R and Add Notes
Reviewed what was needed in the dataprep section after some basic cleaning. It is clear that the location variable is not useful as a character string without a lot of regular expressions. We should consider factorizing it. We also should provide dollar amounts when possible, fix misclassified points, and form key-value pairs while trying to answer the big picture question: What do we want to gain from this?