Amortization Complete, Begin App Development
Completed the amortization calculation with all 30 year mortgage rates from 1971-04-02 through the 2022-04-21. This was based on a fixed value of 100k house with no down payment and the interest rate given as the APR. We should make the date, amount of loan, down payment, and term length variable. We should also predict the interest rate and overlay income growth or average income with the same date selections as the amortization to see how wide the gap is in housing affordability over time. These income values should be adjusted for inflation but the mortgage rates do not need to be seasonally adjusted (the significant or the effect on an annual would be negligible for the problems at hand).